# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.SendEmailRequest:
    properties:
      message:
        type: string
      order_id:
        type: string
      recipient:
        type: string
      subject:
        type: string
    required:
    - order_id
    - recipient
    - subject
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Multi-tenant SaaS para vendas e atendimento via WhatsApp
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: IAFarma API
  version: "1.0"
paths:
...
  /orders/send-email:
    post:
      consumes:
      - application/json
      description: Send order details to a customer via email
      parameters:
      - description: Email request data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SendEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Send order details via email
      tags:
      - orders
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"